refactor(desktop): keep selector controls bounded
